\name{EPISbigRR-package}
\alias{EPISbigRR-package}
\docType{package}
\title{
Generalized Ridge Regression for Analysis of Epistatic Interactions
}
\description{
The package fits large-scale (generalized) ridge regression. It extends the previous bigRR package to include models with interactions. 
}
\details{
\tabular{ll}{
Package: \tab bigRR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-05\cr
Depends: \tab R (>= 2.10.0), utils, hglm, DatABEL\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Authors: Lars Ronnegard and Xia Shen\cr
Maintainer: Lars Ronnegard <lrn@du.se>
}
\references{
Shen X, Alam M, Fikse F and Ronnegard L (2013). \bold{A novel generalized ridge regression method for quantitative genetics}. \emph{Genetics}, \bold{193}, 1255-1268.
}

